requirements.txt
greenlet==3.2.4
hupper==1.12.1
Jinja2==3.1.6
MarkupSafe==3.0.2
PasteDeploy==3.1.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
pyramid==2.0.2
pyramid_jinja2==2.10.1
SQLAlchemy==2.0.43
translationstring==1.4
typing_extensions==4.15.0
venusian==3.1.1
waitress==3.0.2
WebOb==1.8.9
wheel==0.45.1
zope.deprecation==6.0
zope.interface==8.0
zope.sqlalchemy==3.1
